+++ src/main/java/com/daveestar/bettervanilla/Main.java
@@ -33,6 +33,7 @@ import com.daveestar.bettervanilla.events.RightClickCropHarvest;
 import com.daveestar.bettervanilla.events.DoubleDoorSync;
 import com.daveestar.bettervanilla.events.ChestSort;
 import com.daveestar.bettervanilla.events.CropProtection;
+import com.daveestar.bettervanilla.events.ItemRestock;
 import com.daveestar.bettervanilla.events.SignColors;
 import com.daveestar.bettervanilla.events.VanishEvents;
 import com.daveestar.bettervanilla.events.ModerationEvents;
@@ -155,6 +156,7 @@ public class Main extends JavaPlugin {
     manager.registerEvents(new RightClickCropHarvest(), this);
     manager.registerEvents(new DoubleDoorSync(), this);
     manager.registerEvents(new ChestSort(), this);
+    manager.registerEvents(new ItemRestock(), this);
     manager.registerEvents(new VeinMiningChopping(), this);
     manager.registerEvents(new SignColors(), this);
     manager.registerEvents(new VanishEvents(), this);