+++ src/main/java/com/daveestar/bettervanilla/enums/Permissions.java
@@ -23,7 +23,8 @@ public enum Permissions {
   CHESTSORT("bettervanilla.chestsort", Category.PLAYER),
   DOUBLE_DOOR("bettervanilla.doubledoor", Category.PLAYER),
   VEINMINER("bettervanilla.veinminer", Category.PLAYER),
-  VEINCHOPPER("bettervanilla.veinchopper", Category.PLAYER);
+  VEINCHOPPER("bettervanilla.veinchopper", Category.PLAYER),
+  ITEM_RESTOCK("bettervanilla.itemrestock", Category.PLAYER);
 
   public enum Category {
     PLAYER,